Record "Any" architecture in index
The Arduino library specification allows library authors to specify compatibility with any board architecture by using
the `*` value in the `architectures` field of the library.properties metadata file.

The "Arduino Library list" code translates this special `*` architecture value to the more human friendly "Any" for
display to the site's visitors.

The code previously did not record the "Any" architecture type in the architectures index, which meant it was not
possible to browse the list of libraries which are specified as compatible with any architecture.